+++ apps/backend/src/main/kotlin/ca/uvictrades/trades/matching/port/BuyMarketOrder.kt
@@ -1,10 +1,11 @@
 package ca.uvictrades.trades.matching.port
 
 import java.math.BigDecimal
+import java.math.BigInteger
 
 data class BuyMarketOrder(
 	val stock: String,
-	val quantity: Int,
+	val quantity: BigInteger,
 	val liquidity: BigDecimal,
 	val trader: String
 )
