+++ apps/backend/src/main/kotlin/ca/uvictrades/trades/matching/port/SellLimitOrderResidue.kt
@@ -1,13 +1,14 @@
 package ca.uvictrades.trades.matching.port
 
 import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.Instant
 
 data class SellLimitOrderResidue(
 	val orderId: String,
 	val stockId: String,
-	val matchedQuantity: Int,
-	val remainingQuantity: Int,
+	val matchedQuantity: BigInteger,
+	val remainingQuantity: BigInteger,
 	val pricePerUnit: BigDecimal,
 	val timestamp: Instant,
 	val trader: String
